organization: changed 'common' to more descriptive 'math' and moved/renamed serac_types
