+++ src/CMakeLists.txt
@@ -4,7 +4,7 @@
 #
 # SPDX-License-Identifier: (BSD-3-Clause)
 
-add_subdirectory(common)
+add_subdirectory(math)
 add_subdirectory(solvers)
 add_subdirectory(integrators)
 add_subdirectory(coefficients)
