+++ src/solvers/CMakeLists.txt
@@ -27,7 +27,7 @@ blt_add_library(
     NAME        solvers
     SOURCES     ${solvers_sources}
     HEADERS     ${solvers_headers}
-    DEPENDS_ON  mfem integrators common infrastructure physics_utilities physics_operators axom fmt mpi
+    DEPENDS_ON  mfem integrators math infrastructure physics_utilities physics_operators axom fmt mpi
     )
 
 target_include_directories(solvers PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/..)
