+++ src/solvers/base_solver.hpp
@@ -16,11 +16,11 @@
 #include <map>
 #include <memory>
 
-#include "common/serac_types.hpp"
 #include "mfem.hpp"
 #include "solvers/physics_utilities/boundary_condition.hpp"
 #include "solvers/physics_utilities/equation_solver.hpp"
 #include "solvers/physics_utilities/finite_element_state.hpp"
+#include "solvers/physics_utilities/solver_config.hpp"
 
 namespace serac {
