+++ src/solvers/physics_utilities/equation_solver.hpp
@@ -17,8 +17,8 @@
 #include <optional>
 #include <variant>
 
-#include "common/serac_types.hpp"
 #include "mfem.hpp"
+#include "solvers/physics_utilities/solver_config.hpp"
 
 namespace serac {
