+++ tests/serac_dtor.cpp
@@ -10,7 +10,7 @@
 #include <fstream>
 #include <memory>
 
-#include "common/mesh_utils.hpp"
+#include "math/mesh_utils.hpp"
 #include "mfem.hpp"
 #include "serac_config.hpp"
 #include "solvers/thermal_solver.hpp"
