+++ tests/serac_linelastic_solver.cpp
@@ -8,7 +8,7 @@
 
 #include <fstream>
 
-#include "common/mesh_utils.hpp"
+#include "math/mesh_utils.hpp"
 #include "mfem.hpp"
 #include "serac_config.hpp"
 #include "solvers/elasticity_solver.hpp"
