+++ tests/serac_quasistatic_solver.cpp
@@ -9,7 +9,7 @@
 #include <fstream>
 
 #include "coefficients/loading_functions.hpp"
-#include "common/mesh_utils.hpp"
+#include "math/mesh_utils.hpp"
 #include "mfem.hpp"
 #include "serac_config.hpp"
 #include "solvers/nonlinear_solid_solver.hpp"
